clusternator-json.js migrated into project-fs API
