+++ src/api/0.1/cli/aws.js
@@ -1,10 +1,8 @@
 'use strict';
 
 const cn = require('../js/js-api');
-const cmn = require('../common');
 const fs = require('../project-fs/fs');
-
-const clusternatorJson = cmn.src('clusternator-json');
+const clusternatorJson = require('../project-fs/clusternator-json');
 
 module.exports = {
   listProjects: cn.listProjects,
