+++ src/api/0.1/cli/stdio-inheritors.js
@@ -1,15 +1,14 @@
 'use strict';
 
 const Q = require('q');
-const fs = require('fs');
+const fs = require('../project-fs/fs');
 const path = require('path');
 const mkdirp = Q.nfbind(require('mkdirp'));
 
 const logSsh = require('./log-ssh');
-const cn = require('../js/js-api');
+const clusternatorJson = require('../project-fs/clusternator-json');
 
 const cmn = require('../common');
-const clusternatorJson = cmn.src('clusternator-json');
 const util = cmn.src('util');
 const constants = cmn.src('constants');
 const sshKey = cmn.src('cli-wrappers', 'ssh-keygen');
@@ -26,7 +25,7 @@ module.exports = {
 
 function newSshKey(name) {
   return Q.all([
-      clusternatorJson.findProjectRoot(),
+      fs.findProjectRoot(),
       clusternatorJson.get()])
     .then((results) => {
       const publicPath =
