+++ src/api/0.1/project-fs/init.js
@@ -1,0 +1,29 @@
+const fs = require('./fs');
+
+module.exports = initProject;
+
+/**
+ * @param {string} root
+ * @param {{ deploymentsDir: string, clusternatorDir: string,
+ projectId: string, backend: string, tld: string, circleCi: boolean }} options
+ * @param skipNetwork
+ * @returns {Request|Promise.<T>|*}
+ */
+function initProject(root, options, skipNetwork) {
+  var dDir = options.deploymentsDir,
+    cDir = options.clusternatorDir,
+    projectId = options.projectId,
+    dockerType = options.backend;
+
+  return Q
+    .allSettled([
+      deploymentsFs.init(dDir, projectId, options.ports),
+      scriptsFs.init(cDir, options.tld),
+      scriptsFs.initOptional(options, root),
+      dockerFs.init(cDir, dockerType)])
+    .then(() => {
+      if (skipNetwork) {
+        util.info('Network Resources *NOT* Checked');
+      }
+    });
+}
